bist: fixed USB hub reset issue
The USB type C needs to be reset if a USB 3.0 is used on the connector, otherwise this might be detected as a USB 2.0. The original design would reset all USB connections which lead to a locking of the Microchip USB Hub after a while.
This FIX only resets the the Type C USB connection if a USB 2.0 is detected on the bus. This happens only when the "usb-bist,5" is called; as this is the command regarding the USB 3.0 on the Type C connector test